+++ src/TgBotLib/Bot.php
@@ -2135,6 +2135,28 @@
         }
 
         /**
+         * Use this method to change the bot's description, which is shown in the chat with the bot if the chat is empty.
+         * Returns True on success.
+         *
+         * @param string|null $description New bot description; 0-512 characters. Pass an empty string to remove the dedicated description for the given language.
+         * @param string|null $language_code A two-letter ISO 639-1 language code. If empty, the description will be applied to all users for whose language there is no dedicated description.
+         * @param array $options
+         * @return bool
+         * @throws TelegramException
+         * @link https://core.telegram.org/bots/api#setmydescription
+         * @noinspection PhpUnused
+         */
+        public function setMyDescription(?string $description=null, ?string $language_code=null, array $options=[]): bool
+        {
+            $this->sendRequest('setMyDescription', array_merge([
+                'description' => $description,
+                'language_code' => $language_code
+            ], $options));
+
+            return true;
+        }
+
+        /**
          * Use this method to change the bot's menu button in a private chat, or the default menu button.
          * Returns True on success.
          *